cloud_functions/budget_guardian/requirements.txt
google-cloud-firestore>=2.11.0
google-cloud-scheduler>=2.7.0
google-cloud-functions>=1.8.0
